cp2k_creator: Added workFolder kwarg as an alternative to folderPath
This is to make it consistent with CalcMethodFactoryBase. Both these attributes point to the same value, which is changed by changing either. Upon initialisation if EITHER is set to a non-None value then both are. This behaviour wont hold for the create function, but you should never be passing both to create with different values anyway