binned_res: Sorted the betweenVals option for integrating over bins
We now use bin centre by default to decide whether to include a bin in the sum. This should now always lead to the sum adding to total when breaking down into a sum of parts. Previous problem was, for example, binEdges=[0,1,2]  with betweenVals=[0,1.1] and [1.1,1.9] would only include the first bin, which is generally not what we'd want.